requirements.txt
fastapi
uvicorn
motor
python-dotenv
pydantic
langgraph
langchain-core
langchain-openai
langchain-anthropic
langchain-google-genai
langchain-groq
langchain-nvidia-ai-endpoints
pymongo
langsmith
uvicorn[standard]
websockets==10.4
SQLAlchemy>=1.4
python-dotenv
msal
requests
bs4
python-multipart
pytest
pytest-asyncio
schedule>=1.1.0
python-dateutil
hdbscan
backoff